Janky trying to normalize clientId / controllerId
MLAPI player spawn (not done yet)
Janky trying to work in Network Client
Make sure TODOs are warnings

+++ Assets/Scripts/Core/Input/InputManager.cs
@@ -90,7 +90,7 @@ namespace pdxpartyparrot.Core.Input
         }
 
         [CanBeNull]
-        public Gamepad GetGamepad(short playerControllerId)
+        public Gamepad GetGamepad(int playerControllerId)
         {
             int gamepadCount = GetGamepadCount();
             if(playerControllerId >= gamepadCount) {
